# Perf improvements: Reduce amount of theory solving