This PR significantly improves the performance of the SMT solver with two main changes:

* Before doing full theory solving, a subset of the model is attempted, to more quickly get theory clauses
* Equalities of booleans are clausified for the SAT solver so that the theory solver won't spend time refuting them

We can see a noticeable improvement in the GitHub CI times:

* current `main`: https://github.com/CurtisFenner/shiru-ts/actions/runs/4821794438/jobs/8588156874
    > Slowest: verify_tests.duplicated-assert-is-fast took 2772 ms
    > Done in 4.96s.
* this PR: https://github.com/CurtisFenner/shiru-ts/actions/runs/4842439831/jobs/8629394094?pr=9
    > Slowest: verify_tests.duplicated-assert-is-fast took 1776 ms
    > Done in 4.19s.
    * Slowest has -36% runtime
    * Overall has -15% runtime